producer/requirements.txt

flask==3.0.3
kafka-python==2.0.2
gunicorn==21.2.0
